factor out Config class (file)
Config is a static class with globally available configuration data.
It is responsible for reading and checking the command line arguments
and setting up the environment where the following parts of the
script can work in.

NOTE: The script doesn't work ATM, as I'm rebuilding
it from the previous functionality to a new - and
hopefully better - architecture.

This commit additionally contains a lot of
auto-formatting from my new IDE